Clang-tidy fix WIP

+++ tasks/melnik_i_min_neigh_diff_vec/tests/functional/main.cpp
@@ -34,19 +34,19 @@ class MelnikIMinNeighDiffVecRunFuncTestsProcesses : public ppc::util::BaseRunFun
       return std::get<0>(output_data) == -1 && std::get<1>(output_data) == -1;
     }
 
-    // Вычисляем ожидаемый результат: минимальная разница и минимальный индекс при равенстве
-    int expected_idx = 0;
+    long unsigned int expected_idx = 0;
     int min_diff = std::abs(vector[1] - vector[0]);
-    for (size_t i = 1; i < vector.size() - 1; ++i) {
+    for (long unsigned int i = 1; i < vector.size() - 1; ++i) {
       int curr_diff = std::abs(vector[i + 1] - vector[i]);
-      if (curr_diff < min_diff || (curr_diff == min_diff && static_cast<int>(i) < expected_idx)) {
+      if (curr_diff < min_diff || (curr_diff == min_diff && i < expected_idx)) {
         min_diff = curr_diff;
-        expected_idx = static_cast<int>(i);
+        expected_idx = i;
       }
     }
 
     auto [first, second] = output_data;
-    return first == expected_idx && second == expected_idx + 1;
+    return static_cast<unsigned long int>(first) == expected_idx &&
+           static_cast<unsigned long int>(second) == expected_idx + 1;
   }
 
   InType GetTestInputData() final {
